Fixed kernel transaction failures on closing RelationshipTraversalCursor
It turns out that in Neo4j 4.3 (not earlier) if we break from loops that iterate over relationships, the cursors are not cleaned up.
We will report this as a bug to the kernel developers, but we can work around it by never using `break` or `continue label` from within `getRelationship` loops.

Note that using `unsupported.dbms.debug.trace_cursors=true` can give a red-herring in that it points to the first such cursor created, not the one that actually caused the issue, which makes it very hard to track down.